+++ wondrous-app/components/Common/BoardsActivity/styles.tsx
@@ -1,22 +1,6 @@
 import styled from 'styled-components';
 import { BoardFiltersContainer } from 'components/Common/BoardFilters/styles';
 
-export const BoardsActivityWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0 auto;
-  width: 100%;
-  height: 100%;
-  margin-top: 22px;
-  & > div {
-    margin-left: 18px;
-    :first-child {
-      margin: 0;
-    }
-  }
-`;
-
 export const BoardsActivityInlineViewWrapper = styled.div`
   width: 100%;
   display: flex;
