+++ wondrous-app/components/Dashboard/boards/styles.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { BoardsSubheaderWrapper } from 'components/organization/wrapper/styles';
 
 export const BoardsContainer = styled.div`
   margin-top: 42px;
@@ -22,7 +21,12 @@ export const Wrapper = styled.div`
   margin-top: 30px;
 `;
 
-export const BoardsActivityWrapper = styled(BoardsSubheaderWrapper)`
+export const BoardsActivityWrapper = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-row-gap: 20px;
+  align-items: center;
+  margin-bottom: 20px;
   && {
     grid-template-columns: 1fr;
     .FiltersTrigger-button {
